requirements.txt
appnope==0.1.0
aprslib==0.6.46
APScheduler==3.6.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
defusedxml==0.5.0
dnspython==1.16.0
ecdsa==0.13
entrypoints==0.3
Flask==1.1.1
Flask-APScheduler==1.11.0
flickr-api==0.7.3
httplib2==0.13.1
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.1
oauth2==1.9.0.post1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
pylint==2.3.1
pymongo==3.8.0
pyOpenSSL==19.0.0
pyrsistent==0.15.4
PySocks==1.7.0
python-dateutil==2.8.0
python-oauth2==1.1.1
pytz==2019.2
pyzmq==18.0.2
qtconsole==4.5.5
requests==2.22.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
